# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Djamel E Shop API - Local Marketplace Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List all listings with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: city
          in: query
          schema:
            type: string
        - name: wilaya
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: listingType
          in: query
          schema:
            type: string
            enum: [product, service, store]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedListings"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      security: [{ clerkAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/nearby:
    get:
      operationId: getNearbyListings
      tags: [listings]
      summary: Get listings near user's location (wilaya/city)
      parameters:
        - name: wilaya
          in: query
          required: true
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Nearby listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /listings/featured:
    get:
      operationId: getFeaturedListings
      tags: [listings]
      summary: Get featured/trending listings
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Featured listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /listings/{listingId}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a single listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
    put:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      security: [{ clerkAuth: [] }]
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingBody"
      responses:
        "200":
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      security: [{ clerkAuth: [] }]
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /listings/my:
    get:
      operationId: getMyListings
      tags: [listings]
      summary: Get current user's listings
      security: [{ clerkAuth: [] }]
      responses:
        "200":
          description: My listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"

  # ── Orders ────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics for the current user
      security: [{ clerkAuth: [] }]
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed for the current user
      security: [{ clerkAuth: [] }]
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        wilaya:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          default: "DZ"
        preferredLanguage:
          type: string
          enum: [ar, fr, en]
          default: ar
        isStore:
          type: boolean
          default: false
        storeName:
          type: string
          nullable: true
        storeDescription:
          type: string
          nullable: true
        preferences:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, clerkUserId, displayName, country, preferredLanguage, isStore, createdAt]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameFr:
          type: string
        nameEn:
          type: string
        icon:
          type: string
          nullable: true
        slug:
          type: string
      required: [id, nameAr, nameFr, nameEn, slug]
    Listing:
      type: object
      properties:
        id:
          type: integer
        sellerUserId:
          type: string
        sellerProfile:
          $ref: "#/components/schemas/Profile"
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        titleAr:
          type: string
        titleFr:
          type: string
          nullable: true
        titleEn:
          type: string
          nullable: true
        descriptionAr:
          type: string
        descriptionFr:
          type: string
          nullable: true
        descriptionEn:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        currency:
          type: string
          default: DZD
        listingType:
          type: string
          enum: [product, service, store]
        condition:
          type: string
          enum: [new, used, refurbished]
          nullable: true
        images:
          type: array
          items:
            type: string
        wilaya:
          type: string
        city:
          type: string
          nullable: true
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        viewCount:
          type: integer
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, sellerUserId, titleAr, descriptionAr, wilaya, listingType, isActive, isFeatured, viewCount, reviewCount, createdAt]
    CreateListingBody:
      type: object
      properties:
        categoryId:
          type: integer
        titleAr:
          type: string
        titleFr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: string
        descriptionFr:
          type: string
        descriptionEn:
          type: string
        price:
          type: number
        currency:
          type: string
          default: DZD
        listingType:
          type: string
          enum: [product, service, store]
        condition:
          type: string
          enum: [new, used, refurbished]
        images:
          type: array
          items:
            type: string
        wilaya:
          type: string
        city:
          type: string
      required: [titleAr, descriptionAr, wilaya, listingType]
    UpdateListingBody:
      type: object
      properties:
        categoryId:
          type: integer
        titleAr:
          type: string
        titleFr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: string
        descriptionFr:
          type: string
        descriptionEn:
          type: string
        price:
          type: number
        currency:
          type: string
        listingType:
          type: string
          enum: [product, service, store]
        condition:
          type: string
          enum: [new, used, refurbished]
        images:
          type: array
          items:
            type: string
        wilaya:
          type: string
        city:
          type: string
        isActive:
          type: boolean
    PaginatedListings:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [listings, total, page, totalPages]
...
    DashboardStats:
      type: object
      properties:
        totalListings:
          type: integer
        activeListings:
          type: integer
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        totalSales:
          type: integer
        totalPurchases:
          type: integer
        unreadMessages:
          type: integer
        averageRating:
          type: number
          nullable: true
      required: [totalListings, activeListings, totalOrders, pendingOrders, totalSales, totalPurchases, unreadMessages]
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [new_order, order_status_change, new_message, new_review, listing_view]
        titleAr:
          type: string
        titleFr:
          type: string
          nullable: true
        referenceId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, type, titleAr, createdAt]